[mypyc] Refactor: extract build_type_map etc. from genops
